Restructure by extracting the navbar
