+++ src/pages/Facility/settings/layout.tsx
@@ -15,8 +15,7 @@ import UpdateDevice from "@/pages/Facility/settings/devices/UpdateDevice";
 import { GeneralSettings } from "./general/general";
 import LocationList from "./locations/LocationList";
 import LocationView from "./locations/LocationView";
-import FacilityOrganizationUsers from "./organizations/FacilityOrganizationUsers";
-import FacilityOrganizationView from "./organizations/FacilityOrganizationView";
+import FacilityOrganizationList from "./organizations/FacilityOrganizationList";
 
 interface SettingsLayoutProps {
   facilityId: string;
@@ -24,12 +23,13 @@ interface SettingsLayoutProps {
 
 const getRoutes = (facilityId: string) => ({
   "/general": () => <GeneralSettings facilityId={facilityId} />,
-  "/departments": () => <FacilityOrganizationView facilityId={facilityId} />,
-  "/departments/:id": ({ id }: { id: string }) => (
-    <FacilityOrganizationView facilityId={facilityId} id={id} />
-  ),
-  "/departments/:id/users": ({ id }: { id: string }) => (
-    <FacilityOrganizationUsers facilityId={facilityId} id={id} />
+  "/departments": () => <FacilityOrganizationList facilityId={facilityId} />,
+  "/departments/:id/:tab": ({ id, tab }: { id: string; tab: string }) => (
+    <FacilityOrganizationList
+      facilityId={facilityId}
+      organizationId={id}
+      currentTab={tab}
+    />
   ),
   "/locations": () => <LocationList facilityId={facilityId} />,
   "/location/:id": ({ id }: { id: string }) => (
